backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
PyMuPDF==1.24.0
pytesseract==0.3.13
Pillow==10.4.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic[email]
motor==3.5.0
pymongo==4.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
